inicio da tela de encaminhamento da galeria.
