+++ frontend/src/pages/guardian/gallery-dashboard-guardian.html
@@ -4,14 +4,14 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="icon" href="/frontend/assets/logo-icon.svg">
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css" rel="stylesheet"
         integrity="sha384-T3c6CoIi6uLrA9TneNEoa7RxnatzjcDSCmG1MXxSR1GAsXEV/Dwwykc2MPK8M2HN" crossorigin="anonymous">
     <link rel="stylesheet"
         href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@24,400,0,0" />
-    <link rel="stylesheet" href="/frontend/src/styles/dashboard-guardian.css">
-    <script type="module" src="../../../build/pages/guardian/dashboard.js" defer></script>
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL" crossorigin="anonymous"></script>
-    <title>Dashboard Responsável</title>
+    <link rel="stylesheet" href="/frontend/src/styles/gallery-guardian.css">
+    <script type="module" src="../../../build/pages/guardian/gallery-dashboard-guardian.js" defer></script>
+    <title>Galeria</title>
 </head>
 
 <body>
@@ -35,7 +35,7 @@
                         <a class="nav-link" href="schedule-guardian.html">Agenda</a>
                     </li>
                     <li class="nav-item">
-                        <a class="nav-link" aria-disabled="true">Galeria</a>
+                        <a class="nav-link disabled" aria-disabled="true">Galeria</a>
                     </li>
                     <li class="nav-item">
                         <a class="nav-link disabled" aria-disabled="true">Chat</a>
@@ -44,34 +44,17 @@
             </div>
         </div>
     </nav>
-    <main>
-        <div class="summary">
-            <div class="summary-children">
-                <h5>Meus filhos</h5>
-                <div class="children" id="children">
-                    <div class="best-children-one" id="best-children-one">
-                    </div>
-                    <div class="best-children-two">""</div>
-                    <div class="best-children-three">""</div>
-                </div>
-            </div>
-            <div class="summary-notification">
-                <h5>Notificações</h5>
-                <div class="notification">
-                    
-                </div>
-            </div>
-        </div>
-        <div class="summary-last-info">
-            <div class="last-info-children-evaluation">
-                <h5>Desempenho dos filhos</h5>
-            </div>
-            <div class="last-info-calendar">
-                <h5>Calendário</h5>
-            </div>
+    <main class="">
+        <div class="d-flex flex-direction-row m-5 flex-wrap justify-content-center gap-4" id="box"></div>
+        <div id="myModal" class="modal">
+            <span class="close">&times;</span>
+            <img class="modal-content" id="modalImage">
+            <div id="caption" class="d-flex justify-content-center"></div>
         </div>
     </main>
+
     <footer>Desenvolvido por Esther Lucena, Gabriel Morez, Guilherme Pires e Larissa Wandervegen.</footer>
 </body>
 
 </html>
+
