--- frontend/src/pages/guardian/gallery-guardian.html
@@ -1,56 +1,0 @@
-<!DOCTYPE html>
-<html lang="pt-br">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="icon" href="/frontend/assets/logo-icon.svg">
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css" rel="stylesheet"
-        integrity="sha384-T3c6CoIi6uLrA9TneNEoa7RxnatzjcDSCmG1MXxSR1GAsXEV/Dwwykc2MPK8M2HN" crossorigin="anonymous">
-    <link rel="stylesheet"
-        href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@24,400,0,0" />
-    <link rel="stylesheet" href="/frontend/src/styles/login-page-style-guardian.css">
-    <script type="module" src="../../../build/pages/guardian/gallery-guardian.js" defer></script>
-    <title>Login</title>
-    <style>
-        .modal {
-            display: none;
-            position: fixed;
-            z-index: 1;
-            padding-top: 50px;
-            left: 0;
-            top: 0;
-            width: 100%;
-            height: 100%;
-            overflow: auto;
-            background-color: rgb(0, 0, 0);
-            background-color: rgba(0, 0, 0, 0.9);
-        }
-
-        .modal-content {
-            margin: auto;
-            display: block;
-            width: 80%;
-            max-width: 700px;
-        }
-    </style>
-</head>
-<header>
-    <img class="header-logo" src="/frontend/assets/logo-icon-horizontal-pink.svg" alt="Logo Pequeno Vínculo">
-</header>
-
-<body>
-    <main class="">
-        <div class="d-flex flex-direction-row m-5 flex-wrap justify-content-center gap-4" id="box"></div>
-        <div id="myModal" class="modal">
-            <span class="close">&times;</span>
-            <img class="modal-content" id="modalImage">
-            <div id="caption" class="d-flex justify-content-center"></div>
-        </div>
-    </main>
-
-    <footer>Desenvolvido por Esther Lucena, Gabriel Morez, Guilherme Pires e Larissa Wandervegen.</footer>
-</body>
-
-</html>
-
